Move ChatViewController extensions to separate file
