+++ NextcloudTalk/ChatViewController.swift
@@ -1634,15 +1634,3 @@ import UIKit
         self.dismissKeyboard(true)
     }
 }
-
-extension Notification.Name {
-    static let NCChatViewControllerReplyPrivatelyNotification = Notification.Name(rawValue: "NCChatViewControllerReplyPrivatelyNotification")
-    static let NCChatViewControllerForwardNotification = Notification.Name(rawValue: "NCChatViewControllerForwardNotification")
-    static let NCChatViewControllerTalkToUserNotification = Notification.Name(rawValue: "NCChatViewControllerTalkToUserNotification")
-}
-
-@objc extension NSNotification {
-    public static let NCChatViewControllerReplyPrivatelyNotification = Notification.Name.NCChatViewControllerReplyPrivatelyNotification
-    public static let NCChatViewControllerForwardNotification = Notification.Name.NCChatViewControllerForwardNotification
-    public static let NCChatViewControllerTalkToUserNotification = Notification.Name.NCChatViewControllerTalkToUserNotification
-}
